bugfix + minor refactor: updated synthesis_subroutines.py/blur_patch_mask
___ BugFix:
blend into patch was only using the blured mask and not the distance transform due to a bug introduced when optimizing the code. CV distance transform won't work for U8 when using L2, only for L1 I believe, so I couldn't re-use the already allocated arrays.
This might be a problem for node users, since when this updates rolls to main repo it can change the nodes behavior without a new node version being generated.
Should place a reminder to adjust the decencies at the main repo to reference a concrete bmquilting version instead of just the repo.

___ Refactor:
Removed vignette generation from blur_patch_mask to allow to re-use the function with diff. vignettes.
Will likely re-use it for the circular patches.