Re-implemented quantized dilation using cv2.dilate as a maximum filter for subsets of radii; this results in a more granular output.
__ Notes regarding this implementation __

The for loop also changed to accommodate the new approach, and now the radii are selected in the interval  ] r_i-1, r_i ]   by selecting the radii equal or lower to r_i (in opposition to the prior solution) while already processed radii from prior quant layers (r_i-1<) are tracked with a mask.

OpenCV’s dilate is a generalization of maximum_filter, so the scipy method commented out in the prior commit (relative to point 4) could have been avoided. This allows to keep good performance as openCV’s implementation is optimized and avoid an extra dependency in the project.

Finally, asked LLM to optimize the code and generate the method documentation. Seems fine post revision and after testing, and appears to be slightly faster than the last solution.

__ REMINDER __
I’ve not yet explored using a weight mask when patch searching in the main solution (only in the radial, not yet exposed variant).
Now that I’m limiting the seam path search area with respect to the potential blur radius size, it could be relevant to use weights to penalize pixels close to the overlap border, as these will likely have to be ignored.
This should not be a priority, but it might be something interesting to try out later, so it might be a good idea to create an issue or some other explicit reminder for it somewhere.