when merging new patch seams use the maximum instead for of addition.
unsure which is truly preferable.
As one section is iterated multiple times it risks loosing semantic coherence so addition seems adequate; however, when using vignette addition creates hard transitions between certain patches "seams" due to the overlapping area (not really the seams but the edges of the patch where the vignette is stronger).

Could also choose between the two behaviors, only using max when using vignette, or allowing it to be chosen via the gen configs.

Unsure... will leave it using the maximum until some rationale that justify a different approach comes up.