"fixed" version zero; separated find_patch_vx into two separate functions; updated parallel procedure explanation comment in generate.py.
note:
jena2020 solution is only kept due to the alternative path finding method not requiring pyastar2d lib; the error computation functions are not currently being used.
probably a good idea to separate the version arg into two different arguments, and have the pyastar2d usage be explicitly defined by the user; the hidden behavior override would be acceptable in an application context, but now that the code starts to resemble a proper lib, separated from the comfyui env, it shouldn't obfuscate the functions behavior.