Major refactor: synthesis functions directly fetch from target patch location instead of fetching adjacent blocks; rest of the generation logic was updated accordingly.
Important Dev Note:
Due to the simpler solution fetching the overlap sections directly from the texture the texture initiation starts with zeroes instead of empty in order to prevent NaN or inf values which would be problematic when multiplying with the min-cut mask.

Note to self:
Spent way too much time checking seamless versions because it appeared H Seam patching was not working as intended. Post highlighting the last patch it seems the visual seam was actually from the patch itself, so I either had a bad arg combo, or got unlucky with the testing texture. It is very likely correct, that is: working similarly to prior implementation.
For now, I shouldn't revisit the seamless solution. I can eventually revisit it once circular patching is completed and opt to use the circular approach instead to patch the seam.